fix(APISIMP-SNAPPINNED-PAGECAP): prevent CWE-190 int overflow in skip arithmetic
Use long skip = (long) page * pageSize in SnapshotPinnedReadRest to
prevent integer overflow at large page/pageSize combinations (CWE-190).
Update SnapshotService.listDataObjectAppIdsPage and SnapshotDAO.findDataObjectAppIds
to accept long skip; update test matchers to eq(0L/2L/500L) and anyLong().

+++ backend/src/main/java/de/dlr/shepard/v2/snapshot/resources/SnapshotPinnedReadRest.java
@@ -157,7 +157,7 @@ public class SnapshotPinnedReadRest {
 
     // DB-side SKIP/LIMIT: count + fetch only the requested window
     long total = snapshotService.countDataObjectAppIds(snapshot);
-    int skip = page * pageSize;
+    long skip = (long) page * pageSize;
     List<String> paged = snapshotService.listDataObjectAppIdsPage(snapshot, skip, pageSize);
 
     return Response.ok(new SnapshotDataObjectsIO(snapshot, paged, (int) total, page, pageSize))
